+++ test/DefaultCommitManagerTest.sol
@@ -85,12 +85,12 @@ contract DefaultCommitManagerTest is Test, BattleHelper {
         // Alice commits
         vm.startPrank(ALICE);
         uint8 moveIndex = SWITCH_MOVE_INDEX;
-        bytes32 moveHash = keccak256(abi.encodePacked(moveIndex, uint96(0), uint16(0)));
+        bytes32 moveHash = keccak256(abi.encodePacked(moveIndex, uint104(0), uint16(0)));
         commitManager.commitMove(battleKey, moveHash);
 
         // Alice tries to reveal
         vm.expectRevert(DefaultCommitManager.NotYetRevealed.selector);
-        commitManager.revealMove(battleKey, moveIndex, uint96(0), uint16(0), false);
+        commitManager.revealMove(battleKey, moveIndex, uint104(0), uint16(0), false);
     }
 
     function test_RevealBeforeSelfCommit() public {
@@ -108,13 +108,13 @@ contract DefaultCommitManagerTest is Test, BattleHelper {
         // Alice's turn again to move
         vm.startPrank(ALICE);
         vm.expectRevert(DefaultCommitManager.RevealBeforeSelfCommit.selector);
-        commitManager.revealMove(battleKey, NO_OP_MOVE_INDEX, uint96(0), 0, false);
+        commitManager.revealMove(battleKey, NO_OP_MOVE_INDEX, uint104(0), 0, false);
     }
 
     function test_BattleNotYetStarted() public {
         vm.startPrank(ALICE);
         vm.expectRevert(DefaultCommitManager.BattleNotYetStarted.selector);
-        commitManager.revealMove(bytes32(0), NO_OP_MOVE_INDEX, uint96(0), 0, false);
+        commitManager.revealMove(bytes32(0), NO_OP_MOVE_INDEX, uint104(0), 0, false);
         vm.startPrank(BOB);
         vm.expectRevert(DefaultCommitManager.BattleNotYetStarted.selector);
         commitManager.commitMove(bytes32(0), bytes32(0));
@@ -127,7 +127,7 @@ contract DefaultCommitManagerTest is Test, BattleHelper {
         engine.end(battleKey);
         vm.startPrank(ALICE);
         vm.expectRevert(DefaultCommitManager.BattleAlreadyComplete.selector);
-        commitManager.revealMove(battleKey, NO_OP_MOVE_INDEX, uint96(0), 0, false);
+        commitManager.revealMove(battleKey, NO_OP_MOVE_INDEX, uint104(0), 0, false);
         vm.startPrank(BOB);
         vm.expectRevert(DefaultCommitManager.BattleAlreadyComplete.selector);
         commitManager.commitMove(battleKey, bytes32(0));
@@ -157,7 +157,7 @@ contract DefaultCommitManagerTest is Test, BattleHelper {
         vm.startPrank(ALICE);
         commitManager.commitMove(battleKey, bytes32("1"));
         vm.startPrank(BOB);
-        commitManager.revealMove(battleKey, SWITCH_MOVE_INDEX, uint96(0), uint16(0), false);
+        commitManager.revealMove(battleKey, SWITCH_MOVE_INDEX, uint104(0), uint16(0), false);
         vm.warp(TIMEOUT * validator.PREV_TURN_MULTIPLIER() + 1);
         engine.end(battleKey);
         assertEq(engine.getWinner(battleKey), BOB);
